test(product): move the submission-stamp selector test to its own file
session-intent.test.ts sits at its allowlisted 606-line cap; the new
lastPromptSubmittedAtMs coverage lives in session-intent-selectors.test.ts.
